+++ src/pages/CRFdetail/components/Forms.js
@@ -933,7 +933,6 @@ class FirstDiagnoseForm_5 extends React.Component {
   handleSubmit = e => {
     e.preventDefault()
     this.props.form.validateFields((err, values) => {
-      console.log(values)
       if (!err) {
         //重构clinical_symptoms
         const clinical_symptoms = {}
@@ -1528,7 +1527,15 @@ class FirstDiagnoseTable_6 extends React.Component {
             'diagnose_method[targetedtherapy]_other': ""
           }
         }
-        
+
+        if (values.diagnose_existence === 2) {
+          for (let type of ['biopsy_method_other', 'biopsy_type_other', 'genetic_specimen_other', 'tmb_other']) {
+            if (!values[type]) {
+              values[type] = ''
+            }
+          }
+        }
+
         const { dispatch } = this.props
         const sample_id = window.location.pathname.split('/')[4]
         dispatch({
@@ -1796,10 +1803,10 @@ class FirstDiagnoseTable_6 extends React.Component {
                             initialValue: record.biopsy_type
                           })(
                             <Radio.Group onChange={e => this.handleStateChange('biopsy_type', e)}>
-                              <Radio value='0'>无</Radio>
-                              <Radio value='1'>与第1次活检病理类型一致</Radio>
-                              <Radio value='2'>与第1次活检病理类型不一致
-                                {biopsy_type === '2' || (biopsy_type === '' && record.biopsy_type === '2')
+                              <Radio value='无'>无</Radio>
+                              <Radio value='与第1次活检病理类型一致'>与第1次活检病理类型一致</Radio>
+                              <Radio value='与第1次活检病理类型不一致'>与第1次活检病理类型不一致
+                                {biopsy_type === '与第1次活检病理类型不一致' || (biopsy_type === '' && record.biopsy_type === '与第1次活检病理类型不一致')
                                   ?
                                   <div style={{ display: 'inline-block' }}>
                                     {getFieldDecorator('biopsy_type_other', {
@@ -1819,8 +1826,8 @@ class FirstDiagnoseTable_6 extends React.Component {
                               <Radio value='无'>无</Radio>
                               <Radio value='外周血'>外周血</Radio>
                               <Radio value='原发灶组织'>原发灶组织</Radio>
-                              <Radio value='其他'>
-                                转移灶组织{genetic_specimen === '其他' || (genetic_specimen === '' && record.genetic_specimen === '其他')
+                              <Radio value='转移灶组织'>
+                                转移灶组织{genetic_specimen === '转移灶组织' || (genetic_specimen === '' && record.genetic_specimen === '转移灶组织')
                                   ?
                                   <div style={{ display: 'inline-block' }}>
                                     {getFieldDecorator('genetic_specimen_other', {
@@ -1983,7 +1990,7 @@ class FirstDiagnoseTable_6 extends React.Component {
                       </>
                       : null
                   }
-                  <Form.Item label={labelAfter + '开始时间：'}>
+                  <Form.Item label={labelAfter + '开始时间'}>
                     {getFieldDecorator('start_time', {
                       initialValue: record.start_time ? moment(record.start_time, 'YYYY-MM-DD') : null
                     })(
@@ -2074,3 +2081,618 @@ class FirstDiagnoseTable_6 extends React.Component {
 }
 
 export const FirstDiagnoseTable6 = connect(mapStateToProps)(Form.create()(FirstDiagnoseTable_6))
+
+class FirstDiagnoseForm_7 extends React.Component {
+
+  constructor(props) {
+    super(props)
+    this.state = {
+      biopsy_method: '',
+      tumor_pathological_type: '',
+      genetic_testing_specimen: '',
+      tmb: '',
+    }
+  }
+
+  handleStateChange = (type, { target: { value } }) => {
+    this.setState({ [type]: value })
+  }
+
+  handleSubmit = e => {
+    e.preventDefault()
+    this.props.form.validateFields((err, values) => {
+      console.log(values)
+      if (!err) {
+        //重构time
+        if (values.time) values.time = values.time.format('YYYY-MM-DD')
+
+        const { dispatch } = this.props
+        const sample_id = window.location.pathname.split('/')[4]
+        dispatch({
+          type: 'crf_first_diagnose/modifyLabInspection',
+          payload: { sample_id, body: values }
+        }).then(() => dispatch({
+          type: 'crf_first_diagnose/fetchLabInspection',
+          payload: { sample_id }
+        }))
+      }
+    })
+  }
+
+  render() {
+    const { getFieldDecorator } = this.props.form
+    const { lab_inspection } = this.props.crf_first_diagnose
+    const submitLoading = this.props.loading.effects['crf_first_diagnose/modifyLabInspection']
+
+    return (
+      <Form {...formItemLayout} onSubmit={this.handleSubmit}>
+        <Form.Item label={'开始时间'}>
+          {getFieldDecorator('time', {
+            initialValue: lab_inspection.time ? moment(lab_inspection.time, 'YYYY-MM-DD') : null
+          })(
+            <DatePicker format={'YYYY-MM-DD'} />
+          )}
+        </Form.Item>
+        <h2>血常规及凝血功能</h2>
+        <Divider className={styles.lab_inspection_divider} />
+        <Form.Item label="Hb(g/L)">
+          {getFieldDecorator('Hb_val', {
+            initialValue: lab_inspection.Hb_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(g/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Hb_rank', {
+            initialValue: lab_inspection.Hb_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="RBC_B(×10¹²/L)">
+          {getFieldDecorator('RBC_B_val', {
+            initialValue: lab_inspection.RBC_B_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(×10¹²/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('RBC_B_rank', {
+            initialValue: lab_inspection.RBC_B_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="WBC(×10⁹/L)">
+          {getFieldDecorator('WBC_B_val', {
+            initialValue: lab_inspection.WBC_B_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入监测值(×10⁹/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('WBC_B_rank', {
+            initialValue: lab_inspection.WBC_B_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="Plt(×10⁹L)">
+          {getFieldDecorator('Plt_val', {
+            initialValue: lab_inspection.Plt_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(×10⁹L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Plt_rank', {
+            initialValue: lab_inspection.Plt_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="PT(S)">
+          {getFieldDecorator('PT_val', {
+            initialValue: lab_inspection.PT_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(S)" />)}
+          临床意义判定：
+          {getFieldDecorator('PT_rank', {
+            initialValue: lab_inspection.PT_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <h2>尿常规</h2>
+        <Divider className={styles.lab_inspection_divider} />
+        <Form.Item label="白细胞(个/HP)">
+          {getFieldDecorator('RBC_P_val', {
+            initialValue: lab_inspection.RBC_P_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(个/HP)" />)}
+          临床意义判定：
+          {getFieldDecorator('RBC_P_rank', {
+            initialValue: lab_inspection.RBC_P_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="红细胞(个/HP)">
+          {getFieldDecorator('WBC_P_val', {
+            initialValue: lab_inspection.WBC_P_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(个/HP)" />)}
+          临床意义判定：
+          {getFieldDecorator('WBC_P_rank', {
+            initialValue: lab_inspection.WBC_P_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="尿蛋白(＋/－)">
+          {getFieldDecorator('PRO_val', {
+            initialValue: lab_inspection.PRO_val
+          })(
+            <Radio.Group style={{ marginRight: 20 }}>
+              <Radio value={1}>+</Radio>
+              <Radio value={2}>-</Radio>
+            </Radio.Group>
+          )}
+          临床意义判定：
+          {getFieldDecorator('PRO_rank', {
+            initialValue: lab_inspection.PRO_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <h2>血生化</h2>
+        <Divider className={styles.lab_inspection_divider} />
+        <Form.Item label="ALT(IU/L)">
+          {getFieldDecorator('ALT_val', {
+            initialValue: lab_inspection.ALT_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(IU/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('ALT_rank', {
+            initialValue: lab_inspection.ALT_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="AST(IU/L)">
+          {getFieldDecorator('AST_val', {
+            initialValue: lab_inspection.AST_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(IU/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('AST_rank', {
+            initialValue: lab_inspection.AST_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="TBIL(umol/1)">
+          {getFieldDecorator('TBIL_val', {
+            initialValue: lab_inspection.TBIL_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入监测值(umol/1)" />)}
+          临床意义判定：
+          {getFieldDecorator('TBIL_rank', {
+            initialValue: lab_inspection.TBIL_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="DBIL(umol/1)">
+          {getFieldDecorator('DBIL_val', {
+            initialValue: lab_inspection.DBIL_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(umol/1)" />)}
+          临床意义判定：
+          {getFieldDecorator('DBIL_rank', {
+            initialValue: lab_inspection.DBIL_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="ALB(g/L)">
+          {getFieldDecorator('ALB_val', {
+            initialValue: lab_inspection.ALB_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(g/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('ALB_rank', {
+            initialValue: lab_inspection.ALB_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="Cr(umol/L)">
+          {getFieldDecorator('Cr_val', {
+            initialValue: lab_inspection.Cr_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(umol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Cr_rank', {
+            initialValue: lab_inspection.Cr_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="BUN(mmol/1)">
+          {getFieldDecorator('BUN_val', {
+            initialValue: lab_inspection.BUN_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/1)" />)}
+          临床意义判定：
+          {getFieldDecorator('BUN_rank', {
+            initialValue: lab_inspection.BUN_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="Glu(mmol/L)">
+          {getFieldDecorator('Glu_val', {
+            initialValue: lab_inspection.Glu_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Glu_rank', {
+            initialValue: lab_inspection.Glu_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="K(mmol/L)">
+          {getFieldDecorator('K_val', {
+            initialValue: lab_inspection.K_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('K_rank', {
+            initialValue: lab_inspection.K_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="Na(mmol/L)">
+          {getFieldDecorator('Na_val', {
+            initialValue: lab_inspection.Na_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Na_rank', {
+            initialValue: lab_inspection.Na_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="Cl(mmol/L)">
+          {getFieldDecorator('Cl_val', {
+            initialValue: lab_inspection.Cl_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('Cl_rank', {
+            initialValue: lab_inspection.Cl_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="P(mmol/L)">
+          {getFieldDecorator('P_val', {
+            initialValue: lab_inspection.P_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(mmol/L)" />)}
+          临床意义判定：
+          {getFieldDecorator('P_rank', {
+            initialValue: lab_inspection.P_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <h2>肿瘤标志物</h2>
+        <Divider className={styles.lab_inspection_divider} />
+        <Form.Item label="CEA(ng/ml)">
+          {getFieldDecorator('CEA_val', {
+            initialValue: lab_inspection.CEA_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(ng/ml)" />)}
+          临床意义判定：
+          {getFieldDecorator('CEA_rank', {
+            initialValue: lab_inspection.CEA_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="SCC(U/ml)">
+          {getFieldDecorator('SCC_val', {
+            initialValue: lab_inspection.SCC_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入检测值(U/ml)" />)}
+          临床意义判定：
+          {getFieldDecorator('SCC_rank', {
+            initialValue: lab_inspection.SCC_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Form.Item label="NSE(u/ml)">
+          {getFieldDecorator('NSE_val', {
+            initialValue: lab_inspection.NSE_val
+          })(<Input type="number" style={{ width: 200, marginRight: 30 }} placeholder="请输入监测值(u/ml)" />)}
+          临床意义判定：
+          {getFieldDecorator('NSE_rank', {
+            initialValue: lab_inspection.NSE_rank
+          })(<Radio.Group>
+            <Radio value={1}>1</Radio>
+            <Radio value={2}>2</Radio>
+            <Radio value={3}>3</Radio>
+            <Radio value={4}>4</Radio>
+          </Radio.Group>)}
+        </Form.Item>
+        <Col offset={4}>
+          <Button
+            htmlType="submit"
+            type="primary"
+            loading={submitLoading}
+          >保存</Button>
+        </Col>
+      </Form>
+    )
+  }
+}
+
+export const FirstDiagnoseForm7 = connect(mapStateToProps)(Form.create()(FirstDiagnoseForm_7))
+
+class FirstDiagnoseTable_8 extends React.Component {
+
+  constructor(props) {
+    super(props)
+    this.state = {
+      record: {},
+      visible: false,
+      method: ''
+    }
+  }
+
+  handleDelete = evaluate_id => {
+    Modal.confirm({
+      title: '请问是否确认删除？',
+      okText: '确定',
+      cancelText: '取消',
+      onOk: () => new Promise(resolve => {
+        const { dispatch } = this.props
+        const sample_id = window.location.pathname.split('/')[4]
+        dispatch({
+          type: 'crf_first_diagnose/deletePhotoEvaluateTable',
+          payload: { sample_id, evaluate_id }
+        }).then(() => {
+          resolve()
+          dispatch({
+            type: 'crf_first_diagnose/fetchPhotoEvaluateTable',
+            payload: { sample_id }
+          })
+        })
+      })
+    })
+  }
+
+  handleSubmit = e => {
+    e.preventDefault()
+    this.props.form.validateFields((err, values) => {
+      if (!err) {
+        const { dispatch } = this.props
+        const { record } = this.state
+        const sample_id = window.location.pathname.split('/')[4]
+
+        //重构时间和其他空项
+        if (values.time) values.time = values.time.format('YYYY-MM-DD')
+        for(let type of ['method','method_other','time']){
+          if (!values[type]) values[type] = ''
+        }
+        if(!values.tumor_long) values.tumor_long = null
+        if(!values.tumor_short) values.tumor_short = null
+
+        values.evaluate_id = record.evaluate_id
+        dispatch({
+          type: 'crf_first_diagnose/modifyPhotoEvaluateTable',
+          payload: { sample_id, body: values }
+        }).then(() => {
+          this.setState({ visible: false })
+          dispatch({
+            type: 'crf_first_diagnose/fetchPhotoEvaluateTable',
+            payload: { sample_id }
+          })
+        })
+      }
+    })
+  }
+
+  handleEditModel = record => {
+    this.setState({ record, visible: true })
+  }
+
+  handleCancel = () => {
+    this.setState({ visible: false })
+  }
+
+  handleStateChange = (type, { target: { value } }) => {
+    this.setState({ [type]: value })
+  }
+
+  render() {
+    const { photo_evaluate_table } = this.props.crf_first_diagnose
+    const tableLoading = this.props.loading.effects['crf_first_diagnose/fetchPhotoEvaluateTable']
+    const submitLoading = this.props.loading.effects['crf_first_diagnose/modifyPhotoEvaluateTable']
+    const { getFieldDecorator } = this.props.form
+    const { record, visible, method } = this.state
+
+    const columns = [{
+      title: '部位',
+      dataIndex: 'part',
+      align: 'center'
+    }, {
+      title: '方法',
+      dataIndex: 'method',
+      align: 'center'
+    }, {
+      title: '肿瘤长径(cm)',
+      dataIndex: 'tumor_long',
+      align: 'center'
+    }, {
+      title: '肿瘤短径(cm)',
+      dataIndex: 'tumor_short',
+      align: 'center'
+    }, {
+      title: '时间',
+      dataIndex: 'time',
+      align: 'center'
+    }, {
+      title: '操作',
+      align: 'center',
+      render: (_, record) => (
+        <>
+          <Button type="primary" size="small">
+            上传</Button>
+          <Button style={{ marginLeft: '10px' }} type="primary" size="small"
+            onClick={() => this.handleEditModel(record)}>
+            编辑</Button>
+          <Button style={{ marginLeft: '10px' }} type="danger" size="small"
+            onClick={() => this.handleDelete(record.evaluate_id)}>
+            删除</Button>
+        </>
+      )
+    }]
+
+    return (
+      <>
+        <Button type="primary" onClick={() => this.handleEditModel({ evaluate_id: '' })}>添加</Button>
+        <Table
+          loading={tableLoading}
+          className={styles.patient_report_table}
+          rowKey={'evaluate_id'}
+          size="small"
+          bordered={true}
+          pagination={false}
+          scroll={{ x: true }}
+          columns={columns}
+          dataSource={photo_evaluate_table}
+        />
+        <Modal
+          title="编辑影像学评估"
+          visible={visible}
+          okText="保存"
+          destroyOnClose
+          onCancel={this.handleCancel}
+          centered
+          footer={null}
+        >
+          <Form labelCol={{ span: 6 }} wrapperCol={{ span: 18 }} onSubmit={this.handleSubmit}>
+            <Form.Item label="部位">
+              {getFieldDecorator('part', {
+                initialValue: record.part
+              })(
+                <Input placeholder="请输入部位" />
+              )}
+            </Form.Item>
+            <Form.Item label="方法">
+              {getFieldDecorator('method', {
+                initialValue: record.method
+              })(
+                <Radio.Group onChange={e => this.handleStateChange('method', e)}>
+                  <Radio value='CT'>CT</Radio>
+                  <Radio value='MRI'>MRI</Radio>
+                  <Radio value='超声'>超声</Radio>
+                  <Radio value='X线平片'>X线平片</Radio>
+                  <Radio value='PET-CT'>PET-CT</Radio>
+                  <Radio value='其他'>
+                    数量（个突变/Mb）{method === '其他' || (method === '' && record.method === '其他')
+                      ?
+                      <div style={{ display: 'inline-block' }}>
+                        {getFieldDecorator('method_other', {
+                          initialValue: record.method_other
+                        })(<Input style={{ width: 200, marginLeft: 15 }} placeholder="请输入其他方法" />)}
+                      </div>
+                      : null}
+                  </Radio>
+                </Radio.Group>
+              )}
+            </Form.Item>
+            <Form.Item label="肿瘤长径(cm)">
+              {getFieldDecorator('tumor_long', {
+                initialValue: record.tumor_long
+              })(
+                <Input type="number" placeholder="请输入肿瘤长径(cm)" />
+              )}
+            </Form.Item>
+            <Form.Item label="肿瘤短径(cm)">
+              {getFieldDecorator('tumor_short', {
+                initialValue: record.tumor_short
+              })(
+                <Input type="number" placeholder="请输入肿瘤短径(cm)" />
+              )}
+            </Form.Item>
+            <Form.Item label="时间">
+              {getFieldDecorator('time', {
+                initialValue: record.time ? moment(record.time, 'YYYY-MM-DD') : null
+              })(
+                <DatePicker format={'YYYY-MM-DD'} />
+              )}
+            </Form.Item>
+            <Row type="flex" justify="center">
+              <Button
+                htmlType="submit"
+                type="primary"
+                loading={submitLoading}
+              >保存</Button>
+              <Button
+                style={{ marginLeft: 20 }}
+                onClick={this.handleCancel}
+              >取消</Button>
+            </Row>
+          </Form>
+        </Modal >
+      </>
+    )
+  }
+}
+
+export const FirstDiagnoseTable8 = connect(mapStateToProps)(Form.create()(FirstDiagnoseTable_8))